unify test logger
